fix conversion warning

+++ marc.c
@@ -463,7 +463,7 @@ int marc_poll_event(marc_context_t ctx, MPMessage* event, size_t* size, struct s
 		break;
 	}
 
-	int bytes;
+	ssize_t bytes;
 	socklen_t socklen = sizeof(struct sockaddr);
 	if ( (bytes=recvfrom(ctx->sd, event, sizeof(MPMessage), 0, from, &socklen)) <= 0 ){
 		return errno;
@@ -472,8 +472,8 @@ int marc_poll_event(marc_context_t ctx, MPMessage* event, size_t* size, struct s
 	event->type = ntohl(event->type);
 
 	/* fill in version field for old MArCd versions */
-	const int legacy_marc = ctx->type == CONTEXT_CLIENT && event->type == MP_LEGACY_AUTH_EVENT && bytes < sizeof(struct MPauth);
-	const int legacy_mp   = ctx->type == CONTEXT_SERVER && event->type == MP_LEGACY_INIT_EVENT && bytes < sizeof(struct MPinitialization);
+	const int legacy_marc = ctx->type == CONTEXT_CLIENT && event->type == MP_LEGACY_AUTH_EVENT && (size_t)bytes < sizeof(struct MPauth);
+	const int legacy_mp   = ctx->type == CONTEXT_SERVER && event->type == MP_LEGACY_INIT_EVENT && (size_t)bytes < sizeof(struct MPinitialization);
 	if ( legacy_marc ){
 		event->type = MP_CONTROL_AUTHORIZE_EVENT;
 		event->auth.version.major = 0;
